I think it's working. anyway. TBeta as an addon, custom ofxCvBlobs and ContourFinder, POCO events, independent app, tui sends out dx, dy, maccel, and some other goodies
git-svn-id: http://nuicode.svnrepository.com/svn/tbeta@13 463ed9da-a5aa-4e33-a7e2-2d3b412cff85

+++ trunk/tbeta/OSX/libs/openFrameworks/app/ofAppRunner.cpp
@@ -285,7 +285,7 @@ void ofSetVerticalSync(bool bSync){
 	//--------------------------------------
 	#ifdef TARGET_OSX
 	//--------------------------------------
-		long sync = bSync == true ? 1 : 0;
+		GLint sync = bSync == true ? 1 : 0;
 		CGLSetParameter (CGLGetCurrentContext(), kCGLCPSwapInterval, &sync);
 	//--------------------------------------
 	#endif
